+++ schema.graphql
@@ -2,19 +2,15 @@ type PaprController @entity {
   id: ID!
   createdAt: Int!
   # TODO watch for update pool address events
-  pool: Pool! @derivedFrom(field: "controller")
+  poolAddress: Bytes!
   underlying: ERC20Token!
   paprToken: ERC20Token!
   token0IsUnderlying: Boolean!
   maxLTV: BigInt!
+  target: BigInt!
   vaults: [Vault!] @derivedFrom(field: "controller")
   allowedCollateral: [AllowedCollateral!]! @derivedFrom(field: "controller")
   targetUpdates: [TargetUpdate!] @derivedFrom(field: "controller")
-}
-
-type Pool @entity {
-  id: ID!
-  controller: PaprController!
 }
 
 type Vault @entity {
